git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene5012@1597118 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/util/fst/ForwardBytesReader.java
@@ -41,7 +41,7 @@ final class ForwardBytesReader extends FST.BytesReader {
   }
 
   @Override
-  public void skipBytes(int count) {
+  public void skipBytes(long count) {
     pos += count;
   }
